Reorganise into single files
