+++ .flake8
@@ -1,4 +1,8 @@
 [flake8]
+# Rule definitions: http://flake8.pycqa.org/en/latest/user/error-codes.html
+# F401: 'identifier' imported but unused
 exclude = venv*,__pycache__
 max-complexity = 10
 max-line-length = 120
+per-file-ignores =
+    **/__init__.py : F401
